Fixed Query issue.

+++ client/monitor.cc
@@ -170,7 +170,7 @@ void
 MonitorView::updatePlot() {
   QPainter painter(&_plot);
   painter.drawPixmap(0, 0, _plot, 0, 1, N_PLOT, N_PLOT_HIST);
-  for (size_t i=1; i<N_PLOT; i++) {
+  for (int i=1; i<N_PLOT; i++) {
     double db = 10*std::log10(_psd[PLOT_OFFSET+i-1])
         - 10*std::log10(N_FFT) - 10*std::log10(N_PSD_SUM);
     db = std::max(DB_MIN, std::min(db, DB_MAX));
